# [Perf] Support topk softmax fused kernel for broader num_experts