# Essential Elements of an Effective PR Description Checklist

- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.

## Purpose
Current `topkGatingSoftmaxKernelLauncher` requires num_experts to be power of 2 and fall back to 2 kernel approach to compute softmax and topK separately. This is quite slow for models that don't follow ^2. 

The main reason for enforcing power of 2 is to make sure each warp can handle **exactly** X rows (tokens) with fixed `MAX_BYTES_PER_LDG=16`. However if we reduce it to 8 bytes, it allows us to support more num_experts variants (multiple of 64). And benchmark shows that speed up is substantial comparing to the default 2-kernel approach. 

Additionally added 512 (power of 2) given I saw huge improvement over baseline. This should benefit upcoming sparse**r** models. 

## Test Plan
1. fused_topk layer latency benchmark - used benchmark_moe.py but only included `fused_topk` function in run() func. 
2. tests/kernels/moe/test_moe.py for testing correctness

## Test Result
lower batch sizes 1-64:   `32.7% - 46.2%` latency reduction
higher batch sizes 128+: `39.4% - 86.7%` latency reduction

| Batch Size | 576-top8 New | 576-top8 Baseline | 576-top1 New | 576-top1 Baseline | 192-top8 New | 192-top8 Baseline | 512-top8 New | 512-top8 Baseline |
|------------|-----------|----------------|-----------|----------------|-----------|----------------|-----------|----------------|
| 1 | 8.77 us | 16.77 us | 5.84 us | 8.77 us | 6.30 us | 12.84 us | 7.06 us | 14.96 us |
| 2 | 7.29 us | 13.60 us | 5.02 us | 7.51 us | 7.95 us | 10.52 us | 8.89 us | 12.17 us |
| 4 | 7.28 us | 14.23 us | 5.17 us | 7.51 us | 6.51 us | 10.58 us | 7.23 us | 12.15 us |
| 8 | 7.27 us | 13.63 us | 5.14 us | 7.60 us | 6.85 us | 10.67 us | 7.24 us | 12.24 us |
| 16 | 7.66 us | 13.77 us | 5.54 us | 8.05 us | 6.57 us | 11.09 us | 7.57 us | 12.64 us |
| 32 | 7.69 us | 14.16 us | 5.21 us | 7.75 us | 6.93 us | 11.14 us | 7.63 us | 12.71 us |
| 64 | 7.89 us | 14.37 us | 5.21 us | 7.81 us | 7.13 us | 11.32 us | 7.83 us | 12.90 us |
| 128 | 7.82 us | 14.40 us | 5.27 us | 7.93 us | 7.04 us | 11.36 us | 7.75 us | 12.95 us |
| 256 | 9.92 us | 19.33 us | 6.72 us | 10.13 us | 7.36 us | 15.01 us | 8.03 us | 17.40 us |
| 512 | 8.41 us | 18.40 us | 5.87 us | 9.22 us | 9.23 us | 13.80 us | 10.23 us | 17.24 us |
| 1024 | 9.30 us | 27.08 us | 6.00 us | 10.82 us | 7.93 us | 19.39 us | 9.07 us | 25.34 us |
| 2048 | 11.87 us | 44.97 us | 6.63 us | 15.57 us | 9.22 us | 30.53 us | 11.52 us | 41.78 us |
| 4096 | 17.03 us | 79.02 us | 8.21 us | 23.75 us | 12.30 us | 51.92 us | 16.24 us | 73.73 us |
| 8192 | 26.60 us | 147.07 us | 10.13 us | 40.49 us | 17.96 us | 93.41 us | 25.31 us | 136.10 us |
| 10240 | 31.54 us | 183.25 us | 11.22 us | 52.10 us | 20.81 us | 113.89 us | 29.81 us | 168.58 us |
| 20480 | 55.63 us | 357.25 us | 18.02 us | 107.09 us | 34.69 us | 218.61 us | 52.06 us | 329.03 us |
| 76800 | 247.73 us | 1703.97 us | 60.19 us | 449.88 us | 149.77 us | 1056.43 us | 176.89 us | 1574.37 us |